authorization, Spring Boot 2.0.2

+++ src/main/java/ru/doublebyte/amznsm/controllers/IndexController.java
@@ -1,6 +1,7 @@
 package ru.doublebyte.amznsm.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -8,6 +9,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import ru.doublebyte.amznsm.services.ItemStorage;
+import ru.doublebyte.amznsm.services.SessionStorage;
 import ru.doublebyte.amznsm.services.StockInfo;
 import ru.doublebyte.amznsm.structs.Stock;
 
@@ -22,21 +24,30 @@ public class IndexController {
 
     private static final String DEFAULT_NAME = "Unknown Item";
 
-    private ItemStorage itemStorage;
-    private StockInfo stockInfo;
+    @Value("${access.password}")
+    private String password;
+
+    private final ItemStorage itemStorage;
+    private final StockInfo stockInfo;
+    private final SessionStorage sessionStorage;
 
     private Map<String, Stock> stocks = new HashMap<>();
 
     @Autowired
-    public IndexController(ItemStorage itemStorage, StockInfo stockInfo) {
+    public IndexController(ItemStorage itemStorage, StockInfo stockInfo, SessionStorage sessionStorage) {
         this.itemStorage = itemStorage;
         this.stockInfo = stockInfo;
+        this.sessionStorage = sessionStorage;
     }
 
     ///////////////////////////////////////////////////////////////////////////
 
     @GetMapping("/")
     public String index(Model model) {
+        if (!sessionStorage.isAuthorized()) {
+            return "redirect:/login";
+        }
+
         Map<String, String> items = itemStorage.getItems();
 
         List<Stock> stockItems = items.entrySet().stream()
@@ -72,6 +83,10 @@ public class IndexController {
 
     @PostMapping("/add")
     public String addItem(@RequestParam("link") String link) {
+        if (!sessionStorage.isAuthorized()) {
+            return "redirect:/login";
+        }
+
         link = link.trim();
 
         if (!link.isEmpty()) {
@@ -83,8 +98,45 @@ public class IndexController {
 
     @PostMapping("/delete")
     public String removeItem(@RequestParam("id") String id) {
+        if (!sessionStorage.isAuthorized()) {
+            return "redirect:/login";
+        }
+
         itemStorage.deleteItem(id);
+
         return "redirect:/";
+    }
+
+    ///////////////////////////////////////////////////////////////////////////
+
+    @GetMapping("/login")
+    public String getLogin() {
+        if (sessionStorage.isAuthorized()) {
+            return "redirect:/";
+        } else {
+            return "login";
+        }
+    }
+
+    @PostMapping("/login")
+    public String postLogin(
+            @RequestParam("password") String password,
+            Model model
+    ) {
+        if (this.password.equals(password)) {
+            sessionStorage.setAuthorized(true);
+            return "redirect:/";
+        }
+
+        model.addAttribute("failed", true);
+
+        return "login";
+    }
+
+    @GetMapping("/logout")
+    public String logout() {
+        sessionStorage.setAuthorized(false);
+        return "redirect:/login";
     }
 
 }
